+++ mod/primitives/pkg/ssz/testdata/beacon.ssz
