--- mod/storage/pkg/sszdb/tree/doc.go
@@ -1,7 +1,0 @@
-/*
-Package tree provides a Merkle tree view of an SSZ hash root through
-interactions with ferranbt/fastssz.  It should be replaced with the work
-in mod/primitives/pkg/ssz when completed, but it may also be useful to
-keep around for reference and regression testing for a period.
-*/
-package tree
