--- mod/storage/pkg/sszdb/tree/util.go
@@ -1,34 +1,0 @@
-package tree
-
-import "math/bits"
-
-var zeroBytes = make([]byte, 32)
-
-func parseBitlist(dst, buf []byte) ([]byte, uint64) {
-	msb := uint8(bits.Len8(buf[len(buf)-1])) - 1
-	size := uint64(8*(len(buf)-1) + int(msb))
-
-	dst = append(dst, buf...)
-	dst[len(dst)-1] &^= uint8(1 << msb)
-
-	newLen := len(dst)
-	for i := len(dst) - 1; i >= 0; i-- {
-		if dst[i] != 0x00 {
-			break
-		}
-		newLen = i
-	}
-	res := dst[:newLen]
-	return res, size
-}
-
-func nextPowerOfTwo(v uint64) uint {
-	v--
-	v |= v >> 1
-	v |= v >> 2
-	v |= v >> 4
-	v |= v >> 8
-	v |= v >> 16
-	v++
-	return uint(v)
-}
